+++ src/components/UI/Dropdown.astro
@@ -19,7 +19,7 @@ const { buttonNameDefault, dropdownList, id, localStorageKey } = Astro.props;
     }
   </ul>
 </div>
-<style>
+<style is:global>
   .dropdown {
     box-shadow: 4px 6px 15px 0px var(--background-color);
     border-bottom-left-radius: 10px;
