+++ src/layouts/Layout.astro
@@ -69,6 +69,9 @@ const { title, optionalPreloads } = Astro.props;
     <Footer />
     <style is:global>
       :root {
+        --background-color: black;
+      }
+      [data-theme="alu"] {
         --background-color: #211d29;
         --background-highlight: #35314a;
         --accent-color: #624978;
